v2v: update test cases for v2v testing
