+++ v2v/tests/cfg/linux_vm_check.cfg
@@ -1,6 +1,6 @@
 - linux_vm_check_local:
     vm_type = v2v
-    type = linux_vm_check
+    type = vm_check
     # add your VM name in here
     vms = custom_vm_name
     # add VM user name and password
@@ -9,22 +9,39 @@
 
 - linux_vm_check_remote:
     vm_type = v2v
-    type = linux_vm_check
+    type = vm_check
+    target = ovirt
+    os_type = linux
+    vm_cleanup = yes
     # add your VM name in here
-    vms = custom_vm_name
     # need to get VM's IP on ovirt node not local
     remote_preprocess = yes
-    target = ovirt
     # VM is on ovirt-node not local
     not_preprocess = yes
+    shell_client = ssh
+    shell_port = 22
+    shell_prompt = "^\w:\\.*>\s*$|^\[.*\][\#\$]\s*$"
+    status_test_command = echo $?
+
     # add ovirt-engine url, user name and password
-    ovirt_engine_url = https://xx.xx.xx.xx:8443/api
+    ovirt_engine_url = https://xx.xx.xx.xx/api
     ovirt_engine_user = admin@internal
     ovirt_engine_password = 123456
+
     # add VM user name and password
     username = root
-    password = 123456
+    vm_user = ${username}
+    vm_pwd = 123456
+    vm_name = "CONVERTED_XEN_VM"
+    main_vm = ${vm_name}
+    # Remote XEN host's hostname/ip and password.
+    xen_ip = "XEN.HOSTNAME.EXAMPLE"
+    hostname = ${xen_ip}
+    xen_pwd = "redhat"
+    password = ${xen_pwd}
+
     # add ovirt-node ip, user name and password
     ovirt_node_address = xx.xx.xx.xx
     ovirt_node_user = root
     ovirt_node_password = 123456
+    export_name = FC_export
