+++ v2v/tests/cfg/ovirt.cfg
@@ -1,12 +1,15 @@
 - ovirt:
     vm_type = v2v
     type = ovirt
+    # VM doesn't exist and don't need to deal with it on the local
+    not_preprocess = yes
     # add ovirt-engine url, user name and password
-    ovirt_engine_url = https://xx.xx.xx.xx:8443/api
+    ovirt_engine_url = https://xx.xx.xx.xx/api
     ovirt_engine_user = admin@internal
     ovirt_engine_password = 123456
-    # add your VM name in here
-    vms = custom_vm_name
+    vms = ""
+    # add your imported VM in the export domain of the ovirt
+    main_vm = "IMPORTED_VM"
     export_name = export_storage_name
     storage_name = master_storage_name
     cluster_name = storage_cluster_name
