+++ v2v/tests/src/convert_local_image.py
